fix: apply CodeRabbit findings on PR #10
- notes.md: label the stray kernel line (VPS environment record) and
  complete the Phase 2 verdict sentence
- convert_agent: context-accounting invariant is now an explicit check +
  error log (a bare assert dies silently in the asyncio task and vanishes
  under -O); fail-open publish task kept strongly referenced on self
- analyze_capture: drop redundant int() around round()
- test_vad: fail-open load test exercises the public VadGate.load() path
  via monkeypatch (covers the load→_fail wiring); unused `sent` → `_sent`

37/37 tests still pass.

+++ agent/analyze_capture.py
@@ -273,8 +273,8 @@ def classify_silences(silences, env_in, offset_frames, hop_ms=ENV_HOP_MS,
     gated_min_frac = 0.05
     out = []
     for s, e in silences:
-        i0 = max(0, int(round(s * 1000.0 / hop_ms)) - offset_frames)
-        i1 = max(i0, int(round(e * 1000.0 / hop_ms)) - offset_frames)
+        i0 = max(0, round(s * 1000.0 / hop_ms) - offset_frames)
+        i1 = max(i0, round(e * 1000.0 / hop_ms) - offset_frames)
         seg = env_in[i0:min(i1, len(env_in))]
         frac = float((seg >= ref * thresh_ratio).mean()) if len(seg) and ref > 0 else 0.0
         # map the output span back to the input timeline and overlap with gates
